Refactor frontend structure: Remove Register.js, add AuthContext, implement routing and authentication, create layout components, and enhance user experience with loading states and role-based access control.
