+++ frontend/src/js/components/common/ProtectedRoute.js
@@ -1,0 +1,1 @@
+import { useAuth } from '../../js/context/AuthContext';
