+++ frontend/src/js/components/layout/Layout.js
@@ -1,0 +1,3 @@
+import Header from './js/components/layout/Header';
+import Footer from './js/components/layout/Footer';
+import { useAuth } from '../../js/context/AuthContext';
