+++ frontend/src/js/context/AuthContext.js
@@ -1,0 +1,2 @@
+import React, { createContext, useContext, useState, useEffect } from 'react';
+import axios from 'axios';
