+++ frontend/src/js/pages/MapPage.js
@@ -1,0 +1,1 @@
+import { useAuth } from '../js/context/AuthContext';
